feature/dispatcher : move MessagePersistencePolicy to utils
