+++ hub/dispatcher/src/main/java/com/hubsante/hub/utils/MessagePersistencePolicy.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hubsante.hub.service;
+package com.hubsante.hub.utils;
 
 import static com.hubsante.hub.config.Constants.HEALTH_VHOST_PREFIX;
 import static com.hubsante.hub.config.Constants.NEXSIS_VHOST;
